Refactor Dagger components to use scopes
