+++ app/src/main/java/io/plaidapp/ui/PlaidApplication.kt
@@ -21,7 +21,6 @@ import android.app.Application
 import android.content.Context
 import io.plaidapp.core.dagger.CoreComponent
 import io.plaidapp.core.dagger.DaggerCoreComponent
-import io.plaidapp.core.dagger.MarkdownModule
 
 /**
  * Io and Behold
@@ -31,7 +30,6 @@ class PlaidApplication : Application() {
     private val coreComponent: CoreComponent by lazy {
         DaggerCoreComponent
             .builder()
-            .markdownModule(MarkdownModule(resources.displayMetrics))
             .build()
     }
