+++ core/src/main/java/io/plaidapp/core/dagger/designernews/UpvoteStoryServiceComponent.kt
@@ -18,7 +18,7 @@ package io.plaidapp.core.dagger.designernews
 
 import dagger.Component
 import io.plaidapp.core.dagger.BaseServiceComponent
-import io.plaidapp.core.dagger.CoreDataModule
+import io.plaidapp.core.dagger.CoreComponent
 import io.plaidapp.core.dagger.SharedPreferencesModule
 import io.plaidapp.core.designernews.data.votes.UpvoteStoryService
 
@@ -28,16 +28,17 @@ import io.plaidapp.core.designernews.data.votes.UpvoteStoryService
 @Component(
     modules = [
         UpvoteStoryServiceModule::class,
-        CoreDataModule::class,
-        DesignerNewsDataModule::class
-    ]
+        DesignerNewsDataModule::class,
+        SharedPreferencesModule::class
+    ],
+    dependencies = [CoreComponent::class]
 )
 interface UpvoteStoryServiceComponent : BaseServiceComponent<UpvoteStoryService> {
 
     @Component.Builder
     interface Builder {
-
         fun build(): UpvoteStoryServiceComponent
+        fun coreComponent(component: CoreComponent): Builder
         fun sharedPreferencesModule(module: SharedPreferencesModule): Builder
         fun upvoteServiceModule(module: UpvoteStoryServiceModule): Builder
     }
