+++ designernews/src/main/java/io/plaidapp/designernews/dagger/DataModule.kt
@@ -21,48 +21,22 @@ import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterF
 import dagger.Lazy
 import dagger.Module
 import dagger.Provides
-import io.plaidapp.core.BuildConfig
-import io.plaidapp.core.dagger.CoreDataModule
-import io.plaidapp.core.dagger.SharedPreferencesModule
 import io.plaidapp.core.data.api.DeEnvelopingConverter
-import io.plaidapp.core.designernews.data.login.AuthTokenLocalDataSource
-import io.plaidapp.designernews.data.api.ClientAuthInterceptor
 import io.plaidapp.designernews.data.api.DesignerNewsService
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Qualifier
-import kotlin.annotation.AnnotationRetention.BINARY
-
-@Retention(BINARY)
-@Qualifier
-private annotation class LocalApi
+import javax.inject.Named
 
 /**
  * Dagger module to provide data functionality for DesignerNews.
  */
-@Module(
-    includes = [
-        SharedPreferencesModule::class,
-        CoreDataModule::class
-    ]
-)
+@Module
 class DataModule {
-
-    @LocalApi
-    @Provides
-    fun providePrivateOkHttpClient(
-        upstream: OkHttpClient,
-        tokenHolder: AuthTokenLocalDataSource
-    ): OkHttpClient {
-        return upstream.newBuilder()
-            .addInterceptor(ClientAuthInterceptor(tokenHolder, BuildConfig.DESIGNER_NEWS_CLIENT_ID))
-            .build()
-    }
 
     @Provides
     fun provideDesignerNewsService(
-        @LocalApi client: Lazy<OkHttpClient>,
+        @Named("designerNewsOkHttpClient") client: Lazy<OkHttpClient>,
         gson: Gson
     ): DesignerNewsService {
         return Retrofit.Builder()
