+++ dribbble/src/main/java/io/plaidapp/dribbble/dagger/DribbbleComponent.kt
@@ -19,7 +19,6 @@ package io.plaidapp.dribbble.dagger
 import dagger.Component
 import io.plaidapp.core.dagger.BaseActivityComponent
 import io.plaidapp.core.dagger.CoreComponent
-import io.plaidapp.core.dagger.CoreDataModule
 import io.plaidapp.core.dagger.dribbble.DribbbleDataModule
 import io.plaidapp.dribbble.ui.shot.ShotActivity
 
@@ -37,7 +36,6 @@ interface DribbbleComponent : BaseActivityComponent<ShotActivity> {
 
         fun build(): DribbbleComponent
         fun coreComponent(component: CoreComponent): Builder
-        fun coreDataModule(module: CoreDataModule): Builder
         fun dribbbleModule(module: DribbbleModule): Builder
     }
 }
