+++ dribbble/src/main/java/io/plaidapp/dribbble/dagger/DribbbleModule.kt
@@ -20,7 +20,6 @@ import androidx.lifecycle.ViewModelProviders
 import android.content.Context
 import dagger.Module
 import dagger.Provides
-import io.plaidapp.core.dagger.dribbble.DribbbleDataModule
 import io.plaidapp.core.data.CoroutinesDispatcherProvider
 import io.plaidapp.core.dribbble.data.ShotsRepository
 import io.plaidapp.core.ui.widget.ElasticDragDismissFrameLayout
@@ -34,7 +33,7 @@ import io.plaidapp.dribbble.ui.shot.ShotViewModelFactory
 /**
  * Module providing injections for the :dribbble feature module.
  */
-@Module(includes = [DribbbleDataModule::class])
+@Module
 class DribbbleModule(private val activity: ShotActivity, private val shotId: Long) {
 
     @Provides
